syntax

+++ stressTestMaker.c
@@ -5,7 +5,7 @@
 int main() {
     int i;
     for(i = 0; i < 48000; i++) {
-        int randnum = (int) ((double)srand(time(NULL)) % 97.0 + 32.0);
+        int randnum = (int) (srand(time(NULL)) % 97.0 + 32.0);
         if (randnum == 127) randnum = 9;
         else if (randnum == 128) randnum = 10;
         putchar((char) randnum);
